pyproject.toml
[project]
name = "rsp-windowglare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.13.0",
    "aiodns==4.0.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "attrs==26.1.0",
    "backoff==2.2.1",
    "backports-zstd==1.3.0",
    "bcrypt==5.0.0",
    "boto3==1.42.87",
    "botocore==1.42.87",
    "brotli==1.2.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "colorama==0.4.6",
    "cryptography==46.0.7",
    "diffusers==0.37.1",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.135.3",
    "fastapi-cli==0.0.24",
    "fastapi-cloud-cli==0.16.1",
    "fastar==0.10.0",
    "filelock==3.25.2",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==1.10.1",
    "idna==3.11",
    "imageio==2.37.3",
    "importlib-metadata==9.0.0",
    "inquirerpy==0.3.4",
    "invoke==3.0.3",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jmespath==1.1.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "minio==7.2.20",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "networkx==3.6.1",
    "numpy==2.4.4",
    "packaging==26.0",
    "paramiko==4.0.0",
    "peft==0.18.1",
    "pfzy==0.3.4",
    "pillow==12.2.0",
    "prettytable==3.17.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "psutil==7.2.2",
    "py-cpuinfo==9.0.0",
    "pycares==5.0.1",
    "pycparser==3.0",
    "pycryptodome==3.23.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.2",
    "pydantic-settings==2.13.1",
    "pygments==2.20.0",
    "pynacl==1.6.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.24",
    "pyyaml==6.0.3",
    "regex==2026.4.4",
    "requests==2.33.1",
    "rich==14.3.3",
    "rich-toolkit==0.19.7",
    "rignore==0.7.6",
    "runpod==1.9.0",
    "s3transfer==0.16.0",
    "safetensors==0.7.0",
    "sentry-sdk==2.57.0",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tokenizers==0.22.2",
    "tomli==2.4.1",
    "tomlkit==0.14.0",
    "torch==2.11.0",
    "torchvision>=0.26.0",
    "tqdm==4.67.3",
    "tqdm-loggable==0.4.1",
    "transformers==5.5.3",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.44.0",
    "uvloop==0.22.1",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.6.0",
    "websockets==16.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
]
